Fix libtar build warnings since dev_t type changed
dev_t changed from uint16_t to unit32_t, so update here.

/home/haitao/code/stable/vela-miwear/apps/external/libtar/libtar/lib/libtar.h:288:28: error: 'snprintf' output may be truncated before the last format character [-Werror=format-truncation=]
  288 |  snprintf((oct), (octlen), "%*lo ", (octlen) - 2, (unsigned long)(num))
      |                            ^~~~~~~
lib/encode.c:152:2: note: in expansion of macro 'int_to_oct'
  152 |  int_to_oct(minor(device), t->th_buf.devminor, 8);
      |  ^~~~~~~~~~

+++ lib/encode.c
@@ -148,8 +148,8 @@ th_set_device(TAR *t, dev_t device)
 	printf("th_set_device(): major = %d, minor = %d\n",
 	       major(device), minor(device));
 #endif
-	int_to_oct(major(device), t->th_buf.devmajor, 8);
-	int_to_oct(minor(device), t->th_buf.devminor, 8);
+	int_to_oct((uint16_t)major(device), t->th_buf.devmajor, 8);
+	int_to_oct((uint16_t)minor(device), t->th_buf.devminor, 8);
 }
